Rename classes and edit storage
