+++ src/main/java/CommandParser.java
@@ -1,26 +1,28 @@
 import commands.*;
-import exceptions.DukeException;
+import exceptions.MorganException;
 
 /**
- * This is an InputParser class, which translates user input
+ * This is an CommandParser class, which translates user input
  * into commands.
  */
-public class InputParser {
+public class CommandParser {
     private final static String DELIMITER = "\\s+";
     private final static int COMMAND_INDEX = 0;
+    private final static String DELIMITER_FOUND_ERROR_MESSAGE = "Please "
+            + "avoid using the symbol \"" + Storage.DELIMITER + "\".";
 
     /**
      * Returns the command specified by the user.
      * @param userInput The input string entered by the user.
      * @return The command specified by the user.
-     * @throws DukeException
+     * @throws MorganException
      */
-    public Command getCommand(String userInput) throws DukeException {
+    public Command getCommand(String userInput) throws MorganException {
         String trimmedUserInput = userInput.trim();
         String[] words = trimmedUserInput.split(DELIMITER);
         String commandKeyword = words[COMMAND_INDEX].toLowerCase();
-        if (userInput.contains(StorageParser.DELIMITER)) {
-            throw new DukeException("OOPS!!! Please avoid using the symbol \"" + Storage.DELIMITER + "\".");
+        if (userInput.contains(TaskParser.DELIMITER)) {
+            throw new MorganException(DELIMITER_FOUND_ERROR_MESSAGE);
         }
         switch (commandKeyword) {
         case FindCommand.KEYWORD:
