+++ src/main/java/Storage.java
@@ -1,10 +1,9 @@
 import java.io.File;
 import java.io.IOException;
-import java.io.FileReader;
 import java.io.FileWriter;
 import java.util.Scanner;
 
-import exceptions.DukeException;
+import exceptions.MorganException;
 import tasks.TaskList;
 import tasks.Task;
 
@@ -14,15 +13,15 @@ import tasks.Task;
 public class Storage {
     private static final String DIRECTORY_PATH = "./duke-files";
     private static final String FILE_PATH = DIRECTORY_PATH + "/tasks.txt";
-    private final StorageParser parser = new StorageParser();
+    private final TaskParser parser = new TaskParser();
     private final File file;
-    protected static final String DELIMITER = StorageParser.DELIMITER;
+    protected static final String DELIMITER = TaskParser.DELIMITER;
 
     /**
      * Constructor for Storage.
-     * @throws DukeException
+     * @throws MorganException
      */
-    public Storage() throws DukeException {
+    public Storage() throws MorganException {
         File directory = new File(DIRECTORY_PATH);
         boolean isDirectoryExist = directory.exists();
         if (!isDirectoryExist) {
@@ -32,16 +31,16 @@ public class Storage {
         try {
             this.file.createNewFile();
         } catch (IOException e) {
-            throw new DukeException("OOPS!!! " + e.getMessage());
+            throw new MorganException("OOPS!!! " + e.getMessage());
         }
     }
 
     /**
      * Loads data from specified file.
      * @param taskList The existing list of tasks from the file.
-     * @throws DukeException
+     * @throws MorganException
      */
-    public void load(TaskList taskList) throws DukeException {
+    public void load(TaskList taskList) throws MorganException {
         try {
             Scanner sc = new Scanner(this.file);
             while (sc.hasNextLine()) {
@@ -49,12 +48,12 @@ public class Storage {
                 try {
                     Task task = parser.decode(taskString);
                     taskList.addTask(task);
-                } catch (DukeException e) {
+                } catch (MorganException e) {
                     continue;
                 }
             }
         } catch (IOException e) {
-            throw new DukeException("OOPS!!! " + e.getMessage());
+            throw new MorganException("OOPS!!! " + e.getMessage());
         }
     }
 
@@ -62,9 +61,9 @@ public class Storage {
     /**
      * Saves the list of tasks into a file.
      * @param taskList The list of tasks to be saved.
-     * @throws DukeException
+     * @throws MorganException
      */
-    public void save(TaskList taskList) throws DukeException {
+    public void save(TaskList taskList) throws MorganException {
         try {
             FileWriter fileWriter = new FileWriter(this.file);
             StringBuilder storageString = new StringBuilder();
@@ -79,7 +78,7 @@ public class Storage {
             fileWriter.write(String.valueOf(storageString));
             fileWriter.close();
         } catch (IOException e) {
-            throw new DukeException("OOPS!!! " + e.getMessage());
+            throw new MorganException("OOPS!!! " + e.getMessage());
         }
     }
 }
