+++ src/main/java/exceptions/MorganException.java
