+++ src/test/java/CommandParserTest.java
@@ -2,7 +2,7 @@ import commands.AddDeadlineCommand;
 import commands.AddEventCommand;
 import commands.AddToDoCommand;
 import commands.DeleteCommand;
-import exceptions.DukeException;
+import exceptions.MorganException;
 import org.testng.annotations.Test;
 import tasks.TaskList;
 import tasks.ToDoTask;
@@ -23,7 +23,7 @@ public class InputParserTest {
             String expected = (new DeleteCommand("delete 1")).execute(t1);
             String actual = parser.getCommand("delete 1").execute(t2);
             assertEquals(expected, actual);
-        } catch (DukeException e) {
+        } catch (MorganException e) {
             ui.print(e.getMessage());
         }
     }
@@ -38,7 +38,7 @@ public class InputParserTest {
             String expected = (new DeleteCommand("done 1")).execute(t1);
             String actual = parser.getCommand("done 1").execute(t2);
             assertEquals(expected, actual);
-        } catch (DukeException e) {
+        } catch (MorganException e) {
             ui.print(e.getMessage());
         }
     }
@@ -51,7 +51,7 @@ public class InputParserTest {
             String expected = (new AddToDoCommand("todo do CS2030 quiz")).execute(t1);
             String actual = parser.getCommand("todo do CS2030 quiz").execute(t2);
             assertEquals(expected, actual);
-        } catch (DukeException e) {
+        } catch (MorganException e) {
             ui.print(e.getMessage());
         }
     }
@@ -66,7 +66,7 @@ public class InputParserTest {
             String actual = parser.getCommand("event do CS2030 quiz /at 27-08-2021 18:00")
                     .execute(t2);
             assertEquals(expected, actual);
-        } catch (DukeException e) {
+        } catch (MorganException e) {
             ui.print(e.getMessage());
         }
     }
@@ -81,7 +81,7 @@ public class InputParserTest {
             String actual = parser.getCommand("deadline do CS2030 quiz /by 27-08-2021 18:00")
                     .execute(t2);
             assertEquals(expected, actual);
-        } catch (DukeException e) {
+        } catch (MorganException e) {
             ui.print(e.getMessage());
         }
     }
